Start consolidating shared types in the monorepo
