+++ apps/web/playwright/element-web-test.ts
@@ -20,7 +20,6 @@ import {
     type ToMatchScreenshotOptions,
 } from "@element-hq/element-web-playwright-common";
 
-import type { IConfigOptions } from "../src/IConfigOptions";
 import { type Credentials } from "./plugins/homeserver";
 import { ElementAppPage } from "./pages/ElementAppPage";
 import { Crypto } from "./pages/crypto";
@@ -31,11 +30,6 @@ import { type WorkerOptions, type Services, test as base } from "./services";
 // Enable experimental service worker support
 // See https://playwright.dev/docs/service-workers-experimental#how-to-enable
 process.env["PW_EXPERIMENTAL_SERVICE_WORKER_NETWORK_EVENTS"] = "1";
-
-declare module "@element-hq/element-web-playwright-common" {
-    // Improve the type for the config fixture based on the real type
-    export interface Config extends Omit<IConfigOptions, "default_server_config"> {}
-}
 
 export interface CredentialsWithDisplayName extends Credentials {
     displayName: string;
