+++ packages/shared-types/lib/index.d.ts
@@ -1,0 +1,9 @@
+/*
+Copyright 2026 Element Creations Ltd.
+
+SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
+Please see LICENSE files in the repository root for full details.
+*/
+
+export type * from "./config.json.d.ts";
+export type * from "./utils.d.ts";
