+++ packages/shared-types/package.json
@@ -1,0 +1,19 @@
+{
+    "name": "shared-types",
+    "type": "module",
+    "version": "0.0.0",
+    "private": true,
+    "description": "Shared types for Element Web & Desktop",
+    "author": "element-hq",
+    "license": "SEE LICENSE IN README.md",
+    "types": "./lib/index.d.ts",
+    "files": [
+        "lib"
+    ],
+    "scripts": {
+        "lint:types": "tsc --noEmit"
+    },
+    "devDependencies": {
+        "typescript": "catalog:"
+    }
+}
