+++ packages/shared-types/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "$schema": "http://json.schemastore.org/tsconfig",
+    "compilerOptions": {
+        "rootDir": "./lib",
+        "target": "esnext",
+        "lib": ["es2024"],
+        "strict": true,
+        "types": [],
+        "allowImportingTsExtensions": true,
+        "noEmit": true,
+        "skipLibCheck": true
+    },
+    "include": ["lib"]
+}
